+++ src/main/java/com/locawork/webapi/service/IJobService.java
@@ -10,8 +10,8 @@ import java.util.Optional;
 public interface IJobService {
 
     List<JobEntity> findAll();
-    void startWork(int applyerId, int startTime);
-    void endWork(int applyerId, int endTime);
+    void startWork(long applyerId, int startTime);
+    void endWork(long applyerId, int endTime);
     List<JobEntity> findMyDoneWork(int userId);
     List<JobEntity> findAvailableJobsWithUserToken(Double longitude, Double latitude, Double distance, Integer userId);
     List<JobEntity> findUpcomingWork(Integer userId);
